requirements/base.txt
appdirs==1.4.3
arrow==0.10.0
beautifulsoup4==4.6.0
bs4==0.0.1
Django==1.8.9
django-contenteditable==0.1
django-debug-toolbar==1.8
django-emoticons==1.2
django-forms-bootstrap==3.0.1
django-tinymce==2.6.0
lxml==3.7.3
numpy==1.13.0
olefile==0.44
packaging==16.8
pandas==0.20.2
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.14.2
six==1.10.0
sqlparse==0.2.3
gunicorn==19.4
whitenoise==3.2
django-cors-headers==2.1.0
